Create good jumping off place for branches that add Servicestack assemblies. Compiles and runs
