+++ GUI/GUI.csproj
@@ -55,6 +55,9 @@
   </ItemGroup>
 
   <ItemGroup>
+    <Content Update="Pages\BaseServicesUserRegistration.cshtml">
+      <Pack>$(IncludeRazorContentInPack)</Pack>
+    </Content>
     <Content Update="Pages\BaseServicesUserData.cshtml">
       <Pack>$(IncludeRazorContentInPack)</Pack>
     </Content>
