+++ GUI/Pages/BaseServices.cshtml
@@ -2,10 +2,6 @@
 @inherits Ace.AceGUI.Pages.BaseServicesCodeBehind
 <h1>AceCommander Base Services</h1>
 
-<hr />
-<h2>Register</h2>
-<button type="button" class="btn btn-default" id="UserRegistration" onclick="@UserRegistration">@labelForUserRegistrationButton</button>
-
 @if (@BaseServicesInitializationRspPayload == null)
 {
   <p>Loading (Calling onInitAsync)</p>
@@ -21,9 +17,7 @@ else
     <p>IsAliveRspPayload = @IsAliveRspPayload</p>
   }
   <hr />
-  <h2>Register</h2>
-  <button type="button" class="btn btn-default" id="UserRegistration" onclick="@UserRegistration">@labelForUserRegistrationButton</button>
-
-  <p>BaseServicesInitializationRspPayload = @BaseServicesInitializationRspPayload</p>
+  <p> add User LoginOrRegisterNewUserComponent here</p>
+  <BaseServicesUserRegistration />
 }
