+++ GUI/Pages/BaseServicesUserRegistration.cshtml
@@ -1,0 +1,16 @@
+
+@inherits Ace.AceGUI.Pages.BaseServicesUserRegistrationCodeBehind
+<h2>User Registration</h2>
+
+<!-- If there is ever a need to populate state on this component, or get information from the Agent before displaing this commponent, add a pair of UserRegistrationInitializationReq/RspDTO.
+    But for no, there is no need to initializae any data in this component before it is displayed -->
+<!-- Display the fields of the Servicestack.Auth.Register DTO -->
+<div>
+  <label for="UserName">@labelForUserName</label>
+  <input type="text" class="form-control" id="UserName" bind="@UserName" placeholder="@placeHolderForUserName" />
+</div>
+<button type="button" class="btn btn-default" id="UserRegistrationSubmitButton" onclick="@UserRegistrationSubmit">@labelForUserRegistrationSubmitButton</button>
+
+<!-- Start region During Development and Debugging -->
+<p>BaseServicesUserRegistrationRspDTO = @BaseServicesUserRegistrationRspDTOExpanded</p>
+<!-- End region During Development and Debugging -->
