+++ GUI/Properties/AssemblyInfo.cs
@@ -1,7 +1,7 @@
 using System.Reflection;
 
 // ATAP.Utilities.BuildTooling.targets will update the build (days since 1/1/2000), and revision (seconds since midnight, UTC, / 2) fields each time a new build occurs
-[assembly:AssemblyFileVersion("1.0.6889.30848")]
+[assembly:AssemblyFileVersion("1.0.6892.32555")]
 // ATAP.Utilities.BuildTooling.targets will update the AssemblyInformationalVersion field each time a new build occurs
 [assembly:AssemblyInformationalVersion("1.0.0")]
 [assembly:AssemblyVersion("1.0.0")]
